+++ custom_components/mydolphin_plus/core/managers/entity_manager.py
@@ -3,16 +3,15 @@ from __future__ import annotations
 import json
 import logging
 import sys
-from typing import Any
 
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity import EntityDescription
 from homeassistant.helpers.entity_registry import EntityRegistry, RegistryEntryDisabler
 
-from ...core.helpers.const import *
-from ...core.helpers.enums import EntityStatus
-from ...core.models.domain_data import DomainData
-from ...core.models.entity_data import EntityData
+from ..helpers.const import *
+from ..helpers.enums import EntityStatus
+from ..models.domain_data import DomainData
+from ..models.entity_data import EntityData
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -22,11 +21,13 @@ class EntityManager:
 
     hass: HomeAssistant
     domain_component_manager: dict[str, DomainData]
+    entities: dict[str, EntityData]
 
     def __init__(self, hass, ha):
         self.hass: HomeAssistant = hass
         self._ha = ha
         self.domain_component_manager: dict[str, DomainData] = {}
+        self.entities = {}
 
     @property
     def entity_registry(self) -> EntityRegistry:
@@ -40,14 +41,14 @@ class EntityManager:
         self.domain_component_manager[domain_data.name] = domain_data
 
     def get_domain_data(self, domain: str) -> DomainData | None:
-        domain_data = self.domain_component_manager.get(domain)
+        domain_data = self.domain_component_manager[domain]
 
         return domain_data
 
     def update(self):
         self.hass.async_create_task(self._async_update())
 
-    async def _handle_created_entities(self, entity_id, entity: EntityData, domain_component):
+    async def _handle_disabled_entity(self, entity_id, entity: EntityData):
         entity_item = self.entity_registry.async_get(entity_id)
 
         if entity_item is not None:
@@ -60,77 +61,93 @@ class EntityManager:
             else:
                 entity.disabled = entity_item.disabled
 
-        entity_component = domain_component(self.hass, entity)
-
+    async def _handle_restored_entity(self, entity_id, component):
         if entity_id is not None:
-            entity_component.entity_id = entity_id
+            component.entity_id = entity_id
             state = self.hass.states.get(entity_id)
 
             if state is not None:
                 restored = state.attributes.get("restored", False)
 
                 if restored:
-                    _LOGGER.info(f"Restored {entity_id} ({entity.name})")
+                    _LOGGER.debug(f"Restored {entity_id} ({component.name})")
 
-        entity.status = EntityStatus.READY
+    async def _async_add_components(self):
+        try:
+            components: dict[str, list] = {}
+            for unique_id in self.entities:
+                entity = self.entities.get(unique_id)
+                domain_manager = self.domain_component_manager.get(entity.domain)
 
-        return entity_component
+                if entity.status == EntityStatus.CREATED and domain_manager is not None:
+                    entity_id = self.entity_registry.async_get_entity_id(
+                        entity.domain, DOMAIN, unique_id
+                    )
+
+                    await self._handle_disabled_entity(entity_id, entity)
+
+                    component = domain_manager.initializer(self.hass, entity)
+
+                    await self._handle_restored_entity(entity_id, component)
+
+                    domain_components = components.get(entity.domain, [])
+                    domain_components.append(component)
+
+                    components[entity.domain] = domain_components
+
+                    entity.status = EntityStatus.READY
+
+            for domain in self.domain_component_manager:
+                domain_manager = self.domain_component_manager.get(domain)
+
+                domain_components = components.get(domain, [])
+                components_count = len(domain_components)
+
+                if components_count > 0:
+                    domain_manager.async_add_devices(domain_components)
+
+                    _LOGGER.info(f"{components_count} {domain} components created")
+
+        except Exception as ex:
+            exc_type, exc_obj, tb = sys.exc_info()
+            line_number = tb.tb_lineno
+
+            _LOGGER.error(
+                f"Failed to add component, "
+                f"Error: {str(ex)}, "
+                f"Line: {line_number}"
+            )
+
+    async def _async_delete_components(self):
+        deleted = 0
+        for unique_id in self.entities:
+            entity = self.entities.get(unique_id)
+
+            if entity.status == EntityStatus.DELETED:
+                entity_id = self.entity_registry.async_get_entity_id(
+                    entity.domain, DOMAIN, unique_id
+                )
+
+                entity_item = self.entity_registry.async_get(entity_id)
+
+                if entity_item is not None:
+                    _LOGGER.info(f"Removed {entity_id} ({entity.name})")
+
+                    self.entity_registry.async_remove(entity_id)
+
+                del self.entities[entity.id]
+
+                deleted += 1
+
+        if deleted > 0:
+            _LOGGER.info(f"{deleted} components deleted")
 
     async def _async_update(self):
         _LOGGER.debug("Starting to update entities")
 
-        last_handled_domain = None
-
         try:
-            for domain in self.domain_component_manager:
-                last_handled_domain = domain
-                domain_data = self.domain_component_manager.get(domain)
-                entities_to_add = []
-                entities_to_delete = {}
-
-                if domain_data is None:
-                    _LOGGER.error(f"Failed to get domain manager for {domain}")
-
-                for name in domain_data.entities:
-                    entity = domain_data.entities.get(name)
-
-                    entity_id = self.entity_registry.async_get_entity_id(
-                        domain, DOMAIN, entity.unique_id
-                    )
-
-                    if entity.status == EntityStatus.CREATED:
-                        entity_component = await self._handle_created_entities(entity_id,
-                                                                               entity,
-                                                                               domain_data.initializer)
-
-                        if entity_component is not None:
-                            entities_to_add.append(entity_component)
-
-                    elif entity.status == EntityStatus.DELETED:
-                        entities_to_delete[entity_id] = entity
-
-                entities_count = len(entities_to_add)
-
-                if entities_count > 0:
-                    domain_data.async_add_devices(entities_to_add)
-
-                    _LOGGER.info(f"{entities_count} {domain} components created")
-
-                deleted_count = len(entities_to_delete.keys())
-                if deleted_count > 0:
-                    for entity_id in entities_to_delete:
-                        entity = entities_to_delete[entity_id]
-
-                        entity_item = self.entity_registry.async_get(entity_id)
-
-                        if entity_item is not None:
-                            _LOGGER.info(f"Removed {entity_id} ({entity.name})")
-
-                            self.entity_registry.async_remove(entity_id)
-
-                        domain_data.delete_entity(entity.name)
-
-                    _LOGGER.info(f"{deleted_count} {domain} components deleted")
+            await self._async_add_components()
+            await self._async_delete_components()
 
         except Exception as ex:
             exc_type, exc_obj, tb = sys.exc_info()
@@ -138,7 +155,6 @@ class EntityManager:
 
             _LOGGER.error(
                 f"Failed to update, "
-                f"Last handled domain: {last_handled_domain}, "
                 f"Error: {str(ex)}, "
                 f"Line: {line_number}"
             )
@@ -148,7 +164,8 @@ class EntityManager:
                       state: str,
                       attributes: dict,
                       device_name: str,
-                      entity_description: EntityDescription | None = None):
+                      entity_description: EntityDescription | None = None,
+                      details: dict | None = None):
         msgs = []
 
         if entity.state != state:
@@ -165,6 +182,12 @@ class EntityManager:
 
         if entity_description is not None and entity.entity_description != entity_description:
             msgs.append(f"Description {str(entity.entity_description)} -> {str(entity_description)}")
+
+        if details is not None and entity.details != details:
+            from_details = self._get_attributes_json(entity.details)
+            to_details = self._get_attributes_json(details)
+
+            msgs.append(f"Details {from_details} -> {to_details}")
 
         modified = len(msgs) > 0
 
@@ -186,17 +209,8 @@ class EntityManager:
 
         return result
 
-    def get(self, domain: str, name: str) -> EntityData | None:
-        domain_data = self.get_domain_data(domain)
-
-        entity = domain_data.get_entity(name)
-
-        return entity
-
-    def _set(self, entity: EntityData, destructors: list[bool] = None):
-        domain_data = self.get_domain_data(entity.domain)
-
-        entity = domain_data.set_entity(entity, destructors)
+    def get(self, unique_id: str) -> EntityData | None:
+        entity = self.entities.get(unique_id)
 
         return entity
 
@@ -207,26 +221,21 @@ class EntityManager:
                    attributes: dict,
                    device_name: str,
                    entity_description: EntityDescription | None,
-                   action: Any = None
+                   details: dict | None = None,
+                   destructors: list[bool] = None
                    ):
 
-        entity_name = entity_description.key
-
-        entity_description.name = entity_name
-
-        domain_data = self.get_domain_data(domain)
-
-        entity = domain_data.get_entity(entity_name)
+        entity = self.entities.get(entity_description.key)
 
         if entity is None:
             entity = EntityData(entry_id, entity_description)
-            entity.domain = domain
             entity.status = EntityStatus.CREATED
+            entity.domain = domain
 
             self._compare_data(entity, state, attributes, device_name)
 
         else:
-            was_modified = self._compare_data(entity, state, attributes, device_name, entity_description)
+            was_modified = self._compare_data(entity, state, attributes, device_name, entity_description, details)
 
             if was_modified:
                 entity.status = EntityStatus.UPDATED
@@ -235,6 +244,20 @@ class EntityManager:
             entity.state = state
             entity.attributes = attributes
             entity.device_name = device_name
-            entity.action = action
+            entity.details = details
 
-        self._set(entity)
+        if destructors is not None and True in destructors:
+            if entity.status == EntityStatus.CREATED:
+                entity = None
+
+            else:
+                entity.status = EntityStatus.DELETED
+
+        if entity is None:
+            _LOGGER.debug(f"{entity_description.name} ignored")
+
+        else:
+            self.entities[entity_description.key] = entity
+
+            if entity.status != EntityStatus.READY:
+                _LOGGER.info(f"{entity.name} ({entity.domain}) {entity.status}, state: {entity.state}")