+++ CHANGELOG.md
@@ -6,6 +6,12 @@
   - `stop` and `pause` are equivalent to `turn_off` (ie stop a cleaning cycle)
   - locate turns the LED on for 2 seconds and off again
 
+- Major refactor to `Core`
+  - all components are now part of the `Core`
+  - Implementation should be done only in API, HA Manager and Configuration Manager
+
+- Remapped vacuum status for each action - turn on, turn off, toggle, start, stop, pause
+
 ## v0.1.0
 - Major refactor of HA Manager, Entity Manager and API (code cleanup)
 - AWS IOT Broker works with asynchronous operations